fix: rules

+++ src/web/controller/asset.ts
@@ -123,6 +123,19 @@ export class AssetController {
       }
 
       const assetId = Number(c.req.param("id"));
+      const asset = await this.service.getAsset(assetId);
+      if (!asset) {
+        return serveNotFound(c);
+      }
+      //only and master role or admin or the owner of the lead
+      if (
+        user.role !== "master" &&
+        user.role !== "owner" &&
+        asset.user_id !== user.id
+      ) {
+        return serveBadRequest(c, ERRORS.NOT_ALLOWED);
+      }
+
       await this.service.deleteAsset(assetId);
 
       return c.json({ message: "Asset deleted successfully" });
